graph/simple: update for int64 IDs

+++ graph/simple/densegraph_test.go
@@ -122,7 +122,7 @@ func TestDenseLists(t *testing.T) {
 	sort.Sort(ordered.ByID(nodes))
 
 	for i, node := range dg.Nodes() {
-		if i != node.ID() {
+		if int64(i) != node.ID() {
 			t.Errorf("Node list doesn't return properly id'd nodes")
 		}
 	}
